pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "support-ticket-triage"
version = "0.1.0"
description = "OpenEnv Support Ticket Triage"
dependencies = [
    "openenv-core",
    "fastapi",
    "pydantic",
    "uvicorn",
    "openai"
]

[project.scripts]
server = "server.app:main"
